Change "public" to "private" in README code example
Please ignore this if I'm mistaken, but reading The JWT Handbook, they refer to using your "private-key or secret" to sign data server-side. This is in line with expectations, and confirms my confusion when reading the code example for usage of this package - you don't specify a public key, but a private one, right?